move dyad_envs.h from core to common as it is used by multiple components
