+++ src/dyad/core/CMakeLists.txt
@@ -1,5 +1,5 @@
 set(DYAD_CORE_SRC ${CMAKE_CURRENT_SOURCE_DIR}/dyad_core.c)
-set(DYAD_CORE_PRIVATE_HEADER ${CMAKE_CURRENT_SOURCE_DIR}/dyad_envs.h ${CMAKE_CURRENT_SOURCE_DIR}/dyad_core.h)
+set(DYAD_CORE_PRIVATE_HEADER ${CMAKE_CURRENT_SOURCE_DIR}/../common/dyad_envs.h ${CMAKE_CURRENT_SOURCE_DIR}/dyad_core.h)
 set(DYAD_CORE_PUBLIC_HEADER)
 
 add_library(${PROJECT_NAME}_core SHARED ${DYAD_CORE_SRC}
